AutoPaperGen/requirements.txt
# LangGraph - 多 Agent 状态机编排
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Qwen API 支持
dashscope>=1.0.0
openai>=1.0.0

# LlamaIndex - RAG 与检索
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-vector-stores-chroma>=0.1.0

# 向量库
chromadb>=0.5.0

# 数据验证
pydantic>=2.0.0

# 模板渲染
jinja2>=3.1.0

# PDF 解析
pymupdf>=1.24.0

# 工具函数
pyyaml>=6.0
regex>=2024.0.0
